+++ cli/cmd/search.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"github.com/apex/log"
 	"github.com/spf13/cobra"
 	"github.com/tarantool/tt/cli/cmdcontext"
 	"github.com/tarantool/tt/cli/modules"
@@ -9,9 +10,11 @@ import (
 )
 
 var (
-	local        bool
-	includeDev   bool
-	includeDebug bool
+	local     bool
+	debug     bool
+	searchCtx = search.SearchCtx{
+		Filter: search.SearchRelease,
+	}
 )
 
 // NewSearchCmd creates search command.
@@ -24,6 +27,14 @@ func NewSearchCmd() *cobra.Command {
 			"tt - tarantool CLI Community Edition\n" +
 			"tarantool - tarantool Community Edition\n" +
 			"tarantool-ee - tarantool Enterprise Edition",
+		Example: `
+# Remote search across all versions of Tarantool Enterprise Edition.
+
+    $ tt search tarantool-ee
+
+# Remote search across all 2.11 debug versions of Tarantool Enterprise Edition.
+
+    $ tt search tarantool-ee --debug --version 2.11`,
 		Run: func(cmd *cobra.Command, args []string) {
 			err := modules.RunCmd(&cmdCtx, cmd.CommandPath(), &modulesInfo,
 				internalSearchModule, args)
@@ -33,10 +44,10 @@ func NewSearchCmd() *cobra.Command {
 	}
 	searchCmd.Flags().BoolVarP(&local, "local-repo", "", false,
 		"search in local files")
-	searchCmd.Flags().BoolVarP(&includeDev, "dev", "", false,
-		"include dev builds of tarantool-ee SDK")
-	searchCmd.Flags().BoolVarP(&includeDebug, "dbg", "", false,
-		"include debug builds of tarantool-ee SDK")
+	searchCmd.Flags().BoolVar(&debug, "debug", debug,
+		"search for debug builds of tarantool-ee SDK")
+	searchCmd.Flags().StringVar(&searchCtx.ReleaseVersion, "version", searchCtx.ReleaseVersion,
+		"specify version")
 	return searchCmd
 }
 
@@ -52,10 +63,16 @@ func internalSearchModule(cmdCtx *cmdcontext.CmdCtx, args []string) error {
 	if len(args) != 1 {
 		return util.NewArgError("incorrect arguments")
 	}
-	searchCtx := search.SearchCtx{Dbg: includeDebug, Dev: includeDev}
 	if local {
+		if debug || (len(searchCtx.ReleaseVersion) > 0) {
+			log.Warnf("--debug and --version options can only be used to" +
+				" search for tarantool-ee packages.")
+		}
 		err = search.SearchVersionsLocal(cmdCtx, cliOpts, args[0])
 	} else {
+		if debug {
+			searchCtx.Filter = search.SearchDebug
+		}
 		err = search.SearchVersions(cmdCtx, searchCtx, cliOpts, args[0])
 	}
 	return err